backend/requirements.txt
aiosqlite==0.20.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
click==8.3.1
colorama==0.4.6
fastapi==0.134.0
GeoAlchemy2==0.18.3
greenlet==3.3.2
h11==0.16.0
idna==3.11
numpy==2.4.2
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.1
shapely==2.1.2
SQLAlchemy==2.0.47
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
requests==2.32.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
